Added BodyFirstLineEmpty rule
Rule that asserts that the 2nd line of the commit message (first line of
the commit message body) is empty.
This is the first multiline rule, i.e. a rule that acts on multiple
lines at once (in this case, the rule just checks the first line of the
multiline body).