Determine files changed in last commit
When no commit message is passed to stdin, gitlint currently determines
the commit message to lint by taking the commit message of the directory
it is currently in. This commit adds support for gitlint to also fetch
the changed files of the last commit. This will be used to create more
complex rules.

Also: added wheel requirement so that we can do distribute a wheel
alongside the source distribution.