Rules now take a gitcontext in validate()
All rules now take a GitContext object as a second parameter in their
validate() method. This context will be used to create more complex
rules that for example also look at the files that have been changed in
a commit.

Also: added wheel requirement so that we can do distribute a wheel
alongside the source distribution.