Adding GitCommit class representing a git commit
Up until now, gitlint only had to notion of a git commit message.
This commit adds the GitCommit class which stores both the commit
message, and meta-info related to the commit (author, date, commit id, etc).

This will allow us to write gitlint rules based on meta-info (e.g.
restrictions on the author email address format).

For now, the commit_msg property of the GitContext class just refers to
the commit_msg of the last commit. This was done as a temporary measure to
reduce the amount of code change in a single commit.

Also made some minor documentation sites updates (not published yet).