Refactoring: rules working on a git commit
In the last commit, we still had the commit_msg property of the GitContext class
refer to the commit_msg of the last commit in order to minimize the amount of
change in a single commit.
This commit continues the refactoring of rules and tests to work on a
commit message instead of on the commit_msg property of a gitcontext object.