Initial support for user-defined rules
This is the initial commit adding support for user-defined rules. It's a
rather large commit, but I decided to keep a few things out to
still try to keep the size manageable.

In particular, there will be follow-up commits adding more tests,
documentation and validation checks as well as general cleanup.

What is present in the current commit:
- Support for the 'extra-path' parameter that specifies where to search
  for user-defined rules
- A new user_rules.py module that implements the business logic for
  finding user-defined rule classes
- Example user-defined rules in examples/my_commit_rules.py and
  examples/my_line_rules.py
- A new DirectoryOption class for specifying directory paths as rule
  options
- ./run_tests.sh now print py.test warnings