Dependency version bumps
Updated all dependencies to the latest version. This brings us to back
to 2016 :-)

The only dependency not getting an update is 'sh' because the 0.12.x
release series seems to be very buggy (performance issues, regressions, etc).

This commit does contain a guard for checking '_ok_code' during sh
invocation as that was an issue with 1.12.x (submitted
https://github.com/amoffat/sh/issues/343). Might as well patch it
already if we ever do go to sh==1.12.x.