Bugfix: ignore precedence
The 'ignore' precedence logic was fundamentally broken. This commit
fixes this.

In order to do this, the ignore business logic moved from LintConfig to
GitLinter.

Before, the LintConfig object already pruned the set of rules as soon as
the ignore attribute was set. The problem with this is that rules didn't
get added back when the ignore config was overwritten by a config option
with higher precedence.

Now, LintConfig only stores the ignore list and ignoring is done at
runtime, during linting.

Also made the following 2 changes are they are related:
 - Simplified GitLint general option parsing.
 - ListOption now also strips whitespace from elements when passed a
   list instead of only stripping whitespace from a csv string.