LintConfigBuilder class for better handling of configuration precedence
This commit adds a new LintConfigBuilder class which acts as a factory
for building LintConfig objects.
The primary reason for this is to fix a number of issues with
configuration precedence. Because of the pervasiveness of LintConfig,
this did require  a lot of refactoring, hence why this commit is quite
substantial.

There will be one more follow-up commit to fix some config precedence
issues with extra-path.