Initial support for debug logging
This commit implements some of the groundwork neccessary to do granular
logging. This will allow developers and users to debug their gitlint
issues more easily when using the --debug flag. We leverage the python
logging module for this.

For now, we only log the commit message sha and the exit code, but a lot
more can be added in the future.