New rule: author-valid-email
This commits adds a new rule (author-valid-email) that validates that the commit
author email address is a valid email address. The rule has a regex option that
allows users to configure specify which email addresses they want to be
considered valid.

Also changed the 'unittest' import back to 'unittest2' to fix a regression with
the integration tests in python 2.6.