Git command execution convenience function
This commit introduces the gitlint.git._git(...) convenience function for
executing git commands.

The main benefit of this is that all error and unicode handling of git commands
is now uniform as it's done in a single place.

In addition, this refactor has the effect of the 'sh' library only being invoked
in a single place in the entire gitlint codebase. This will make it easier to
replace it in the future, for example when we'd consider adding Windows support
(see #20).

This commit also fixes some unit tests related to custom comment chars as
introduced by #34.