Fixes for Debian packaging (#49)
* Add support for python3 for gitlint/tests

Add conditional imports for python3 modules to support testing with
python3. Make dependency for ordereddict conditional for python
versions < 2.7 as it is a python built-in since python version 2.7.

* Remove duplicate line

The exact same line is already executed two lines before and a second
execution will not change the result.

* Change get_sample() and get_expected() to context

Change get_sample() and get_expected() to use context manager to remove
the warnings (ResourceWarning: unclosed file ...) under python3.

* Fix regexp matching in tests

Escape the path in expected exception messages as the assertRaisesRegex
uses the string as a pattern for a re.search and it may contain some
symbols (e.g. '+' or '*') that lets the regexp not match properly on
other build systems.