New IgnoreByTitle rule
Added a new IgnoreByTitle rule that is based on a new more generic rule
construct called ConfigurationRule.

ConfigurationRules can modify gitlint configuration (and therefore
behavior) on a per commit basis.

Added some tests and docs, more docs and integration tests to follow in
additional commits.

This is part of #54 and #57.