0.10.0 release
The 0.10.0 release adds the ability to ignore commits based on their contents,
support for pre-commit, and important fix for running gitlint in CI environments
(such as Jenkins, Gitlab, etc).

Special thanks to @asottile, @bdrung, @pbregener, @torwald-sergesson,
@RykHawthorn, @SteffenKockel and @tommyip for their contributions.

Since it's becoming increasingly hard to support Python 2.6 and 3.3, we'd like
to encourage our users to upgrade their python version to 2.7 or 3.3+. Future
versions of gitlint are likely to drop support for Python 2.6 and 3.3.

Full Changelog:

- New Rule: ignore-by-title allows users to ignore certain commits by matching a
  regex against a commit message title.
- New Rule: ignore-by-body allows users to ignore certain commits by matching a
  regex against a line in a commit message body.
- Gitlint now supports pre-commit.com.
- Gitlint now has a --msg-filename commandline flag that allows you to specify
  the commit message to lint via a file.
- Gitlint will now be silent by default when a specified commit range is empty.
- Gitlint can now be installed on MacOS by brew via the homebrew-devops tap.
  To get the latest version of gitlint, always use pip for installation.
- If all goes well, gitlint will also be available as a package in the
  Ubuntu 18.04 repositories.
- Bugfixes:
  - We fixed a nasty and recurring issue with running gitlint in CI.
    Hopefully that's the end of it :-)
  - Fix for custom git comment characters.

Full Release details in CHANGELOG.md.