Update gitlint and unit tests for windows
* Update config and hooks
* Update lots of tests
* Use codec.open to support reading UTF-8 on windows with py27

All the changes were to do with forward vs backward slashes.
Use os.path.join everywhere it's needed and add some re.escape()
in to deal with the cases where a backslash formatted path gets
put into a regexp. An abspath or two also required to make it all
work on both platforms.

ConfigParser.readfp is deprecated in py3.3 but read_file is not
available in 2.7. Seems readfp does not give a warning in 3.6 so disable
the warning rather than add another if block for 2.7 support