Ignore pylint errors for bad options
When running an old version of pylint with newer options (e.g. to
switch off "new" warnings), pylint would fail. This switches off the
option checking.
This is needed to make pylint work with the full range of Python
versions (and the compatible pylint versions) ranging from Python 2.6
to Python 3.6+.