Make `test_target` less strict with `git`
An expectation of the `test_target` is that `git` prints a specific
error message when run outside a workdir. This message was changed in
one of the last `git` releases, so the test fails.

Make the `test_target` less strict regarding the message content, just
looking for the 'not a git repository' ought to be enough.