Fix: Configured commentchar crashes gitlint
When a git repository has a custom 'core.commentchar' configured, gitlint wouldn
crash. This was caused by gitlint not properly handling the 'success' case where
'git config' returns an exit code of 0.

In addition, gitlint wouldn't respect the  '—target' parameter to determine the
commentchar, instead just executing the git config in the current working
directory instead of the target directory.

This should fix #59