Improved encoding detection on Windows
Gitlint will now respect the LC_ALL, LC_CTYPE and LANG environment variables
on Windows - in line with git's behavior.

The default fallback is UTF-8, this should lead to improved unicode
parsing for the majority of users.

This is a partial fix for #96.

Also includes a minor refactoring of utils.py that makes it easier to unit test
some of the constants that are defined in that module.