Ability to run integration tests without sh
Integration tests can now be run without using the 'sh' library by using
GITLINT_QA_USE_SH_LIB=0. Note that not all tests are currently passing
in this mode.

Eventually we'll entirely replace the 'sh' library with qa/shell.py, so
we can fully support integration tests on Windows.

For now, we will need to keep the 'sh' library around as certain tests
still rely on it for doing more complex interaction with std IO.

This commit also replaces the invocation of POSIX 'touch' and 'rm' commands with
python code equivalents.

Relates to #93