Simple generic caching primitives
A new @cache decorator can now be used together with the PropertyCache class to
do simple property caching on classes.

Modified the GitContext object to use this feature to lazy-load and cache
the git comment character.

This mechanism is only used for caching Git related class properties right now,
but it's generic enough to be used in any class.

There's room to improve on the current design, in particular removing the need
to inherit from the PropertyCache in order to use the @cache decorator - instead
using a global cache.