Bugfix: Allow None value for options
In options, None was converted to the string "None" instead of the python value
None. This made it impossible to actually have optional options.

Also fixed the resulting issue in the Ignore rules which were raising exceptions
as they now by default tried to compile a None regex (as opposed to the "None"
string before which succeeded by was incorrect).