Make requirement pinning optional
gitlint’s pinned requirements make it difficult to use with other
libraries and tools that have newer requirements, and may hold back
important security updates.  With this change, it’s possible to
install gitlint with its requirements either pinned:

    pip install gitlint[pinned]

or unpinned:

    pip install gitlint

At this time, we document only the pinned option because the
maintainer has indicated a preference for supporting that
configuration.

Signed-off-by: Anders Kaseorg <andersk@mit.edu>