🧪🔧 Switch to using `pypa/build` in CI
This patch switches the CI/CD workflows to use the PyPA-endorsed
PEP 517 build front-end called `build`. It is called without
`--sdist` and `--wheel` CLI options, in which case `build` creates
an sdist from the Git checkout but the following wheel build is made
from that sdist tarball as opposed to using the Git checkout.

This has a side effect of smoke-testing that it's possible to build
wheels from the published sdists. And this is the flow that
`pip install` follows when it's requested to install from an sdist.

Ref: https://github.com/jorisroovers/gitlint/issues/460#issuecomment-1462121899