tools/create-test-repo.sh: use subshell
tools/create-test-repo.sh will now enter a subshell
immediately inside the temporary git repo directory, and reconfigure
hatch to be able to run from that that directory.

This avoids developers having to manually type the cd command to enter
the correct temp repo after creation and having to reconfigure hatch.